+++ tests/ngr/dotnet-csharp/DemoProgram.cs
@@ -1,0 +1,17 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace demo
+{
+    public static class DemoProgram
+    {
+        public static async Task Main(string[] args)
+        {
+                Console.WriteLine("Hallo, Räuber Hotzenplotz.");
+        }
+
+    }
+
+}